all sections updated

+++ travel-mate-client/src/components/trips/TripDetails.js
@@ -13,13 +13,17 @@ class TripDetails extends Component {
   }
 
   componentDidMount = () => {
+    debugger;
     this.getSingleTrip();
   };
 
   getSingleTrip = () => {
     const { params } = this.props.match;
     axios
-      .get(`http://localhost:5000/api/trips/${params.id}`)
+      .get(`http://localhost:5000/api/trips/${params.id}`, {
+        withCredentials: true
+      })
+
       .then(responseFromApi => {
         const theTrip = responseFromApi.data;
         // this.setState({ trip: theTrip });
@@ -42,7 +46,9 @@ class TripDetails extends Component {
   deleteTrip = id => {
     const { params } = this.props.match;
     axios
-      .delete(`http://localhost:5000/api/trips/${params.id}`)
+      .delete(`http://localhost:5000/api/trips/${params.id}`, {
+        withCredentials: true
+      })
       .then(responseFromApi => {
         console.log(responseFromApi);
         this.props.history.push("/trips");
@@ -52,17 +58,44 @@ class TripDetails extends Component {
       });
   };
 
+  ownershipCheck = trip => {
+    debugger;
+    if (
+      this.props.loggedInUser &&
+      trip.owner &&
+      trip.owner._id === this.props.loggedInUser._id
+    ) {
+      return (
+        <div>
+          <div>{this.renderEditForm()} </div>
+          <button onClick={() => this.deleteTrip(this.state._id)}>
+            Delete this trip
+          </button>
+        </div>
+      );
+    } else {
+      //debugger;
+      return (
+        <div>
+          <h4>Contact your travel buddy:</h4>
+          {this.state.owner ? (
+            <p>{this.state.owner.email}</p>
+          ) : (
+            <img src="/public/Loading-icon.gif" alt="loading" />
+          )}
+        </div>
+      );
+    }
+  };
+
   render() {
     return (
       <div>
-        <h1>{this.state.destination}</h1>
-        <p>{this.state.startDate}</p>
-        <p>{this.state.endDate}</p>
-        <p>{this.state.pointsOfInterest}</p>
-        <div>{this.renderEditForm()} </div>
-        <button onClick={() => this.deleteTrip(this.state._id)}>
-          Delete trip
-        </button>
+        <h1 class="black">{this.state.destination}</h1>
+        <p>From: {this.state.startDate}</p>
+        <p>To: {this.state.endDate}</p>
+        <p>Attractions: {this.state.pointsOfInterest}</p>
+        <div>{this.ownershipCheck(this.state)}</div>
         <Link to={"/trips"}>Back to trips</Link>
       </div>
     );
